Added debug option and per-task permissions support (Version 1.1)
Users can now configure the plugin to only fire per-player commands on
users with a specific permission.

Extra Logging was split in to extra logging, which notifies when a task
cannot fire for a reason (like no players online), and debug messages,
which trace what is happening under the hood.